moving test etl class to its own file
